Fix Sort Usings in Inline Diff
Invoking Sort Usings on an added using directive while in diff mode was throwing an ArgumentException because textView.GetTextElementSpan() expects to be passed a SnapshotPoint on the top (projection) buffer, not the subject buffer.

Since the snapshotSpan is only needed for the indicatorFactory and it needs to be in the top buffer, pass the caret position in the top buffer instead.

Instead of passing the snapshotSpan.Snapshot to getCurrentDocumentAsync later (which requires the subject buffer), use the buffer from the caretPoint we grabbed earlier, which is in the subject buffer.

Verified that it fixes sort usings in diff.

Fixes #74594